vertex_pipeline/components/component_base/requirements.txt
kfp==1.8.2
lightgbm==3.2.1
gcsfs==2021.6.1
pandas==1.1.5
pandas-gbq==0.14.1
jinja2==3.0.1
requests==2.25.1
google-cloud-logging==2.5.0
google-cloud-aiplatform==1.1.1
google-cloud-storage==1.39.0
google-cloud-bigquery==2.20.0
google-cloud-bigquery-storage==2.4.0
google-cloud-secret-manager==2.5.0
texthero==1.1.0
jsonlines==3.0.0